move TestUtil to util folder
