+++ src/components/SignOut/SignOut.unit.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, fireEvent, cleanup } from "@testing-library/react";
-import { withTestRouter, withMockFirebase } from "../../TestUtil";
+import { withTestRouter, withMockFirebase } from "../../util/TestUtil";
 import { SignOutButton } from ".";
 import text from "./text";
