+++ src/util/TestUtil.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { MemoryRouter } from "react-router-dom";
-import { FirebaseContext } from "./components/Firebase";
-import { AuthUserContext } from "./components/Session";
+import { FirebaseContext } from "../components/Firebase";
+import { AuthUserContext } from "../components/Session";
 
 const withMockFirebase = (mockFirebase, Component) => (
   <FirebaseContext.Provider value={mockFirebase}>
